+++ src/OpenTelemetry.DynamicControl/CHANGELOG.md
@@ -17,4 +17,8 @@
   policy/source ordering via `PolicyKeyComparer`.
   ([#5017](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5017))
 
+* Added ordered, store-change subscriptions to the internal `PolicyStore` via a
+  new `PolicyChangeNotifier`/`PolicyChangeSubscription` pair.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TODO))
+
 For more details, please refer to the [README](README.md).